+++ framework/soa_systems/src/soa_systems_despawn.c
@@ -17,7 +17,7 @@ void soa_get_destination_reached_despawn_slots(
 	for (usize e = 0; e < entity_count; ++e) {
 		const f32 dx = e_destination->x[e] - e_position->x[e];
 		const f32 dy = e_destination->y[e] - e_position->y[e];
-		const f32 distance = sqrt(dx * dx + dy * dy);
+		const f32 distance = sqrtf(dx * dx + dy * dy);
 		const bool is_close = distance < reach_distance;
 		output[count] = (soa_slot_t){ e };
 		count += is_close;
